move funcs and scale scalers to various
