+++ vsscale/__init__.py
@@ -1,7 +1,8 @@
 # ruff: noqa: F401, F403
 
 from .exceptions import *
-from .funcs import *
+from .generic import *
+from .various import *
 from .helpers import *
 from .mask import *
 from .onnx import *
